% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTideCurve.R
\name{PlotTideCurve}
\alias{PlotTideCurve}
\title{Plots the tide curve}
\usage{
PlotTideCurve(data)
}
\arguments{
\item{data}{The results from TideCurve. Warning: The synthesis period must overlap with the analysis period.}
}
\value{
Generates a plot
}
\description{
Plots the tide curve and residuum to give an overview of the fitting results.
}
