% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResCurve.R
\name{ResCurve}
\alias{ResCurve}
\title{Computes the residuum between the observed data and the synthesis}
\usage{
ResCurve(tcData, obsData)
}
\arguments{
\item{tcData}{The results from the TideCurve function. Warning: The synthesis period must overlap with the analysis period.
Must be a data.table object.}

\item{obsData}{The observation data with the columns observation_date, observation_time and height. See attached data for correct formats.}
}
\value{
A list with two data.tables with the joined data input and the computed difference between the observed data and
the synthesis (res)
}
\description{
This function computes the residuum of the computed lunar and solar synthesis
and the observed data
}
