% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tideObservation.R
\docType{data}
\name{tideObservation}
\alias{tideObservation}
\title{Sample file of high and low water times and heights}
\format{
A data frame with 10267 rows and 3 variables
\describe{
  \item{observation_date}{date of observation, character value in "yyyy/mm/dd" format}
  \item{observation_time}{time of observation, character value in "hh:mm:ss" format}
  \item{height}{observed value, numeric value}
}
}
\usage{
tideObservation
}
\description{
A sample dataset containing observation date, time and height of high and low water
}
\keyword{datasets}
